+++ .buildkite/pipeline.yml
@@ -224,10 +224,16 @@ steps:
         build.branch =~ /^master$$/
     timeout_in_minutes: 240
 
-  - label: "SubPackage -- {{matrix}}"
+  - label: "{{matrix.package}} on CUDA {{matrix.cuda}}"
     matrix:
-      - "CUSTATEVEC"
-      - "CUTENSORNET"
+      setup:
+        package:
+          - "CUSTATEVEC"
+          - "CUTENSORNET"
+          - "CUDNN"
+        cuda:
+          - "10.2"
+          - "11.8"
     plugins:
       - JuliaCI/julia#v1:
           version: 1.6
@@ -242,14 +248,14 @@ steps:
       cuda: "11.0"
       cap: "sm_80"
     env:
-      JULIA_CUDA_VERSION: '11.6'
+      JULIA_CUDA_VERSION: "{{matrix.cuda}}"
       JULIA_CUDA_USE_BINARYBUILDER: 'true'
-      PACKAGE: '{{matrix}}'
+      PACKAGE: '{{matrix.package}}'
     if: build.message !~ /\[skip tests\]/ && !build.pull_request.draft
     timeout_in_minutes: 120
     commands: |
       julia -e 'import Pkg; Pkg.develop(; path = pwd())' # CUDA
-      julia -e 'import Pkg; Pkg.develop(; path = joinpath(pwd(), "lib", ENV["PACKAGE"]))'
+      julia -e 'import Pkg; Pkg.develop(; path = joinpath(pwd(), "lib", lowercase(ENV["PACKAGE"])))'
       julia -e 'import Pkg; Pkg.precompile()'
       julia -e 'import Pkg; Pkg.test(ENV["PACKAGE"])'
