+++ lib/cudnn/Project.toml
@@ -1,0 +1,9 @@
+name = "CUDNN"
+uuid = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
+authors = ["Tim Besard <tim.besard@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
