+++ test/runtests.jl
@@ -171,7 +171,6 @@ ENV["CUDA_VISIBLE_DEVICES"] = join(map(pick->"$(pick.mig ? "MIG" : "GPU")-$(pick
 
 # determine tests to skip
 skip_tests = []
-has_cudnn() || push!(skip_tests, "cudnn")
 has_cusolvermg() || push!(skip_tests, "cusolver/multigpu")
 has_nvml() || push!(skip_tests, "nvml")
 if !has_cutensor() || first(picks).cap < v"6.0"
